Improve build portability and timer callback interfaces
Add devbox environment files for local development setup.
Update Makefile to use conditional libusb/rt flags for better cross-platform builds.
Unify timer callback/context types across Linux and Windows implementations.
Fix time formatting in tddevice initialization output.

+++ tddevice.c
@@ -29,7 +29,7 @@ int tddev1_init_operation(td_context_t* context)
 
 	buffer[1] = TDDEV1_CMD_INIT;
 
-	time(&epoc); sprintf((char*)&buffer[2], "%llu", epoc);
+	time(&epoc); sprintf((char*)&buffer[2], "%lld", (long long)epoc);
 
 	if (TdHidSetReport(context->handle, buffer, context->device_type->output_report_size + 1, USB_HID_REPORT_TYPE_FEATURE))
 	{
